Working Log In + Log Out
User is now able to Log in and log out. Password changed from hash value to clear text. Tried to implement the user's name being displayed, but since it was on a separate page I struggled to get it to work. Could look into that when applications and users code is combined